requirements.txt
langchain
langgraph
langchain_community
langchain_core
langchain_groq
langchain_openai
chromadb
streamlit
tavily-python
langchain-google-genai
google-generativeai
python-dotenv
beautifulsoup4
markitdown[all]
pypdf
sentence-transformers
einops
pi-heif
pymupdf